Added enumeration of permutations
If the number of requested permutations is greater than the actual
number of possible permutations for a given matrix size, then the
program enumerates the possible permutations rather than randomly
sampling from the space of permutations. This will usually be quicker
and gives a deterministic result. Enumeration can be forced by setting
perms to 0.