Update _test.py
Update _test.py to accept all square matrices as inputs. Downside is that the function no longer accepts condensed forms.

According to the original paper, Mantel Test can be applied to any pair of square matrices. It is not constrained to be used only for symmetric matrices with 0's on the main diagonal.

References:
Mantel, N. (1967). "The detection of disease clustering and a generalized regression approach". Cancer Research. 27 (2): 209–220

URL: https://aacrjournals.org/cancerres/article/27/2_Part_1/209/476508/The-Detection-of-Disease-Clustering-and-a